Update Hello World Temporal App
