+++ .gitignore
@@ -29,3 +29,7 @@ build
 
 # IntelliJ Idea
 .idea
+
+bin/*
+
+.vscode/*
