+++ app/build.gradle
@@ -1,4 +1,3 @@
-
 plugins {
     // Apply the java plugin to add support for Java
     id 'java'
@@ -13,40 +12,33 @@ java {
 }
 
 repositories {
-    maven {
-        url "https://plugins.gradle.org/m2/"
-    }
-    maven {
-        url "https://oss.sonatype.org/content/repositories/snapshots/"
-    }
     mavenCentral()
 }
 
 // @@@SNIPSTART hello-world-project-template-java-gradle-dependencies
 dependencies {
-    implementation group: 'io.temporal', name: 'temporal-sdk', version: '1.11.0'
+    implementation group: 'io.temporal', name: 'temporal-sdk', version: '1.18.2'
     implementation group: 'ch.qos.logback', name: 'logback-classic', version: '1.2.11'
-    implementation group: 'commons-configuration', name: 'commons-configuration', version: '1.10'
 
-    testImplementation group: 'io.temporal', name: 'temporal-testing', version: '1.11.0'
+    testImplementation group: 'io.temporal', name: 'temporal-testing', version: '1.18.2'
     testImplementation group: 'junit', name: 'junit', version: '4.13.2'
-    testImplementation group: 'org.mockito', name: 'mockito-core', version: '4.5.1'
+    testImplementation group: 'org.mockito', name: 'mockito-core', version: '5.1.1'
 }
 // @@@SNIPEND
 
 application {
     // Define the main class for the application.
-    mainClassName = 'helloworldapp.InitiateHelloWorld'
+    mainClass = 'helloworldapp.InitiateHelloWorld'
 }
 
 // @@@SNIPSTART hello-world-project-template-java-gradle-tasks
 task sayHello(type: JavaExec) {
-    main = 'helloworldapp.InitiateHelloWorld'
+    mainClass = 'helloworldapp.InitiateHelloWorld'
     classpath = sourceSets.main.runtimeClasspath
 }
 
 task startWorker(type: JavaExec) {
-    main = 'helloworldapp.HelloWorldWorker'
+    mainClass = 'helloworldapp.HelloWorldWorker'
     classpath = sourceSets.main.runtimeClasspath
 }
 // @@@SNIPEND
