+++ app/src/main/java/helloworldapp/HelloWorldWorkflow.java
@@ -7,6 +7,11 @@ import io.temporal.workflow.WorkflowMethod;
 @WorkflowInterface
 public interface HelloWorldWorkflow {
 
+    /**
+     * This is the method that is executed when the Workflow Execution is started. The Workflow
+     * Execution completes when this method finishes execution.
+     */
+    
     @WorkflowMethod
     String getGreeting(String name);
 }
