+++ app/src/main/java/helloworldapp/InitiateHelloWorld.java
@@ -12,13 +12,27 @@ public class InitiateHelloWorld {
         WorkflowServiceStubs service = WorkflowServiceStubs.newLocalServiceStubs();
         // WorkflowClient can be used to start, signal, query, cancel, and terminate Workflows.
         WorkflowClient client = WorkflowClient.newInstance(service);
+
+        /*
+         * Set Workflow options such as WorkflowId and Task Queue so the worker knows where to list and which workflows to execute.
+         */
         WorkflowOptions options = WorkflowOptions.newBuilder()
-                .setTaskQueue(Shared.HELLO_WORLD_TASK_QUEUE)
-                .build();
-        // WorkflowStubs enable calls to methods as if the Workflow object is local, but actually perform an RPC.
+                    .setWorkflowId(Shared.HELLO_WORLD_WORKFLOW_ID)
+                    .setTaskQueue(Shared.HELLO_WORLD_TASK_QUEUE)
+                    .build();
+
+        // Create the workflow client stub. It is used to start our workflow execution.
         HelloWorldWorkflow workflow = client.newWorkflowStub(HelloWorldWorkflow.class, options);
-        // Synchronously execute the Workflow and wait for the response.
+
+        /*
+         * Execute our workflow and wait for it to complete. The call to our getGreeting method is
+         * synchronous.
+         * 
+         * Replace the parameter "World" in the call to getGreeting() with your name.
+         */
         String greeting = workflow.getGreeting("World");
+
+        // Display workflow execution results
         System.out.println(greeting);
         System.exit(0);
     }
