+++ app/src/main/java/helloworldapp/Shared.java
@@ -1,0 +1,12 @@
+// @@@SNIPSTART hello-world-project-template-java-shared-constants
+package helloworldapp;
+
+public interface Shared {
+
+    // Define the task queue name
+    final String HELLO_WORLD_TASK_QUEUE = "HELLO_WORLD_TASK_QUEUE";
+
+    // Define the workflow unique id
+    final String HELLO_WORLD_WORKFLOW_ID = "HELLO_WORLD_WORKFLOW";
+}
+// @@@SNIPEND
