--- src/main/java/helloworldapp/Shared.java
@@ -1,8 +1,0 @@
-// @@@SNIPSTART hello-world-project-template-java-shared-constants
-package helloworldapp;
-
-public interface Shared {
-
-    String HELLO_WORLD_TASK_QUEUE = "HELLO_WORLD_TASK_QUEUE";
-}
-// @@@SNIPEND
